Gradle/base/base-spring-app/library/src/main/java/com/example/app/service/MyService.java
package com.example.demo.service;

import org.springframework.boot.context.properties.EnableConfigurationProperties;
import org.springframework.stereotype.Service;


/**
 * @author Evgeniy Skiba on 25.05.2020
 * @project base-sprinng-boot
 */

@Service
@EnableConfigurationProperties(ServiceProperties.class)
public class MyService {

    private final ServiceProperties serviceProperties;

    public MyService(ServiceProperties serviceProperties) {
        this.serviceProperties = serviceProperties;
    }

    public String message() {
        return this.serviceProperties.getMessage();
    }
}